+++ .eslintignore
@@ -1,0 +1,5 @@
+artifacts
+build
+dist
+dist-es
+node_modules
