+++ .eslintrc.js
@@ -1,6 +1,7 @@
 module.exports = {
     env: {
         browser: true,
+        mocha: true,
         node: true,
     },
     extends: ['eslint:recommended'],
