+++ tests/functional/bootstrap.js
@@ -1,4 +1,3 @@
-/* global window, document */
 import React from 'react'; // eslint-disable-line
 import ReactDOM from 'react-dom'; // eslint-disable-line
 import { expect } from 'chai'; // eslint-disable-line
